backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
PyJWT[crypto]>=2.8.0
httpx>=0.26.0
pydantic>=2.0.0
